+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
+
+## [[v2.19.5]](https://github.com/multiversx/mx-sdk-dapp/pull/897)] - 2023-08-29
 - [Fixed Ledger and Walletconnect login with modals](https://github.com/multiversx/mx-sdk-dapp/pull/895)
 -
 ## [[v2.19.4]](https://github.com/multiversx/mx-sdk-dapp/pull/896)] - 2023-08-28